OpenTelemetry/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
click==8.1.8
Deprecated==1.2.18
fastapi==0.115.8
googleapis-common-protos==1.59.1
grpcio==1.70.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.30.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
protobuf==4.25.6
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
thrift==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0
